requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
streamlit==1.44.1
pydantic==2.11.3
pytest==8.3.5
httpx==0.28.1
requests==2.32.3
python-dotenv==1.0.1
